+++ lib/ui/page/holder/main_holder.dart
@@ -4,7 +4,7 @@ import 'package:jjava_flutter/ui/page/holder/home/home_page.dart';
 import 'package:jjava_flutter/ui/page/holder/my_page/my_page_page.dart';
 import 'package:jjava_flutter/ui/page/holder/solved_question/solved_question_page.dart';
 import 'package:jjava_flutter/ui/page/holder/workspace/workspace_page.dart';
-import 'package:jjava_flutter/ui/page/question/select/question_select_page.dart';
+import 'package:jjava_flutter/ui/page/question/select_page/question_select_page.dart';
 
 class MainHolder extends StatefulWidget {
   const MainHolder({super.key, this.initialIndex = 0});
